[refactor] : entity package 생성, MessageEntity 위치 옮김 #17
