+++ whole/src/main/java/model/entity/MessageEntity.java
@@ -1,4 +1,4 @@
-package model;
+package model.entity;
 
 import java.time.LocalDate;
 import java.util.Date;
